created custom accountStatus hook
